common.h: Move pcm.c stuff to pcm.h
