+++ slowrx.c
@@ -22,6 +22,7 @@
 #include "fsk.h"
 #include "gui.h"
 #include "modespec.h"
+#include "pcm.h"
 #include "video.h"
 
 // The thread that listens to VIS headers and calls decoders etc
