+++ vis.c
@@ -8,6 +8,7 @@
 #include "common.h"
 #include "gui.h"
 #include "modespec.h"
+#include "pcm.h"
 
 /* Handle selection of the VIS mode in the GUI */
 static void onGotVis(guchar VIS) {
